LPS-73481 Moving upgrade processes due to new features to next minor release upgrade
