+++ portal-impl/src/com/liferay/portal/upgrade/UpgradeProcess_7_0_4.java
@@ -21,7 +21,6 @@ import com.liferay.portal.upgrade.v7_0_4.UpgradeGroup;
 import com.liferay.portal.upgrade.v7_0_4.UpgradeEmailAddress;
 import com.liferay.portal.upgrade.v7_0_4.UpgradeExpando;
 import com.liferay.portal.upgrade.v7_0_4.UpgradeMBMailingList;
-import com.liferay.portal.upgrade.v7_0_4.UpgradeSchema;
 import com.liferay.portal.upgrade.v7_0_4.UpgradeUser;
 
 /**
@@ -37,8 +36,6 @@ public class UpgradeProcess_7_0_4 extends UpgradeProcess {
 
 	@Override
 	protected void doUpgrade() throws Exception {
-		upgrade(UpgradeSchema.class);
-
 		upgrade(UpgradeContact.class);
 		upgrade(UpgradeGroup.class);
 		upgrade(UpgradeEmailAddress.class);
